I added some basic import functionality back into eof_menu_file_lyrics_import(), and added updated comments for the functions in lc_import.c.
The import completes, but EOF was segfaulting during a call to strcat() at some point when seeking into the chart.  I determined that it was related to some buffer overflow logic that I had corrected in some pre-SVN edits.  After re-adding my edited version of eof_render_lyric_preview(), the problem appears to be fixed.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@7 1b2aedea-9eff-cdc0-263b-b87a75789507