Created an enhanced function to draw the lyrics in the lyric lane: eof_lyric_draw_truncate()
This function is identical to eof_lyric_draw(), except that it takes the lyric number instead of a pointer.  It then examines the next lyric to calculate the best X2 coordinate for the clipping rectangle, guaranteeing that lyrics are truncated.  It also uses this X2 detection to signal to eof_render_vocal_editor_window() when it can stop parsing the lyrics array due to printing an entire lyric outside the clipping rectangle.

It may be worth reviewing this update to make sure the logic is implemented correctly, but if it is, it should remove a considerable amount of processing required to render the vocal editor window.  If there is a way to prevent eof_lyric_draw'ing lyrics that come before the X1 coordinate of the clipping rectangle, that will also probably improve efficiency.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@106 1b2aedea-9eff-cdc0-263b-b87a75789507