Moved ID3 parsing code all into ID3_parse.c, since there is no EOF specific code in it.  This will make it easier for me to maintain with the rest of my converter logic.
Refined the ID3 parsing so that once the ID3 tag is found, all file parsing is kept within the boundaries of the tag, which will minimize the amount of slow file I/O involved with parsing an ID3 tag that is at the end of the file instead of the beginning, or a tag search that yields no results.  Some before and after tests on a set of 355 MP3 files results in the new logic taking about 6.895 seconds compared to 18.35 seconds before, making it nearly three times as fast.  The speed difference for MP3s that store the ID3 tag at the end of the file will probably be ridiculously much more improved.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@206 1b2aedea-9eff-cdc0-263b-b87a75789507