Begins work on issue 127.  The underlying bug is still probably there, which seemed to cause the if...else if...else code segments to execute one of the if blocks and then some of the else block.
I'm trying to implement an error code reporting system, but it's worth your checking that I got the alert() call correct.  If there's an alert window that can display a formatted string (so I don't have to create a string with itoa()), it would probably be better than the way I did it.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@245 1b2aedea-9eff-cdc0-263b-b87a75789507