Adds a new string eof_loaded_ogg_name[] that eof_load_ogg() and eof_load_ogg_quick() update, allowing the filename of the loaded audio file to be tracked reliably.
The waveform data stores the filename of the audio file used to generate it, so if the user loads another OGG file and re-opens the waveform graph using F5, the waveform data is recreated correctly.  I tested loading several separated audio files from a couple Rock Band songs and it appears to work nicely.  I even was able to notice certain things like in the audio for "I Think I'm Paranoid", they mixed the bass drum to be on the left channel.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@384 1b2aedea-9eff-cdc0-263b-b87a75789507