Works on issue 187:
Adds enough checking to mimic the handling of note phrases.  If a note is placed or moved to a position where a drum note is marked as a cymbal, that note is applied the appropriate cymbal marker.  If a note is marked as a cymbal, all drum notes at that location are applied the appropriate cymbal marker.

Reduces the number of calls to perform timing conversion during MIDI import.

Adds the pro drum logic to MIDI import.  This is pretty new logic, so I added a redundant pass through the drum notes once PART DRUMS is parsed, to clear the cymbal flags where necessary.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@505 1b2aedea-9eff-cdc0-263b-b87a75789507