Works on issues 185 and 205: *The proposed project format has been modified to explain how to store lyric phrases and bookmark phrases.  The handling of bookmarks in the load/save logic has been fixed. *The chart that is generated by the new chart logic now works correctly, and stores the pre-defined track names.  These names are saved and loaded in the new project format.
But most importantly, the new save and load logic is in use.  I haven't stress tested it, but I created a 1 second chart and ensured that notes, phrases and lyrics were correctly saved and loaded.  Custom data blocks of 4 bytes with all bits set are in use just until everything is working, just to make it easier to debug the file in a text editor.

Also, some crashing in the grid snap logic that would occur on very short songs (less than 3 beats long) has been corrected.

git-svn-id: http://editor-on-fire.googlecode.com/svn/trunk@589 1b2aedea-9eff-cdc0-263b-b87a75789507